+++ buildSrc/src/main/kotlin/Config.kt
@@ -83,7 +83,6 @@ object Versions {
     const val kotlinCompileTesting = "1.4.2" // https://github.com/tschuchortdev/kotlin-compile-testing
     const val ktlint = "0.45.2" // https://github.com/pinterest/ktlint
     const val ktor = "1.6.8" // https://github.com/ktorio/ktor
-    const val datetime = "0.3.2" // https://github.com/Kotlin/kotlinx-datetime
     const val nexusPublishPlugin = "1.1.0" // https://github.com/gradle-nexus/publish-plugin
     const val okio = "3.0.0" // https://square.github.io/okio/#releases
     const val serialization = "1.3.0" // https://kotlinlang.org/docs/releases.html#release-details